pyproject.toml
[tool.poetry]
name = "project-royex"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
sqlmodel = "^0.0.11"
python-dotenv = "^1.0.0"
fastapi-utils = "^0.2.1"
psycopg2-binary = "^2.9.9"
watchfiles = "^0.21.0"
web3 = "^6.11.3"
numpy = "^1.26.1"


[build-system]
requires = [
    "poetry-core"
]
build-backend = "poetry.core.masonry.api"